Walljump Done, Wallslide Done, PlayerSpriteSheet Center Point fixed
I added so you can spam jump on the same wall. You know have to switch from left to right wall, to get higher
I erased the x velocity for the walljump, since it created so many bugs and withput it, it gives more freedom for the player movement.
And I added the gravity for the wallsliding state again. I kinda like it
And I edited the center of the spritesheet. The Player is now not glitching around anymore when flipping.